move app_register and app_verify_email to SecurityController and the generated templates to security
